This is getting sweeter, but we now have to deal with the elephant in the room for larger DOM structures - that we have to recreate all of it from scratch every time something changes in the model. The question now is whether we can retain the "functional" simplicity of what we have so far while we incorporate a bit of efficiency there.
One thing to observe with what we have so far is that we have the
ability to lazily create children when we need them. We can exploit this
and take the laziness to a different level by lazily creating the entire
structure on demand via a "render" function. Then by keeping track of
what needs to be updated when interactions happen, we can selectively
update only those parts of the DOM tree.

What I mean is that instead of returning a fully constructed node, we
can return a "render" function which when called with the model will
update the DOM node in-place, creating the necessary structures.

To keep track of things that need to be edited upon certain changes to
the model, we introduce a new property of a DOM element called 'dyno' -
which is an array of functions to call to update the DOM tree. As we
render the first DOM, we keep track of these "dynos" and use them to
pudate in-place whenever the model changes.