Thus far, we've only paid attention to creation and updation of the DOM tree. We haven't said anything about how the model is to be updated as events pour in from the DOM tree. There is an easy way to manage this by introducing "commands" that invoke handlers via some kind of a dispatch mechanism.
For our purposes, we can pretend that such a dispatch occus via a
function (message, arguments...) {} kind of function. So an event handler
can be created like this -

    onclick: function (event) { cmd("IncrementCounter", 1); }

The only thing we propose as part of this protocol is that the first argument
be a string identifying the command. The arguments to be command can be
derived from the event. In order to make this simpler, we introduce a
"handler" function which creates such handlers given a cmd and an instruction string.